+++ document-store/src/main/java/org/hypertrace/core/documentstore/mongo/query/transformer/MongoQueryTransformer.java
@@ -2,7 +2,7 @@ package org.hypertrace.core.documentstore.mongo.query.transformer;
 
 import com.google.common.collect.ImmutableList;
 import java.util.List;
-import org.hypertrace.core.documentstore.query.QueryInternal;
+import org.hypertrace.core.documentstore.query.Query;
 import org.hypertrace.core.documentstore.query.transform.QueryTransformer;
 
 public class MongoQueryTransformer {
@@ -13,8 +13,8 @@ public class MongoQueryTransformer {
           .add(new MongoSelectionsUpdatingTransformer())
           .build();
 
-  public static QueryInternal transform(final QueryInternal query) {
-    QueryInternal transformedQuery = query;
+  public static Query transform(final Query query) {
+    Query transformedQuery = query;
 
     for (QueryTransformer transformer : TRANSFORMERS) {
       transformedQuery = transformer.transform(transformedQuery);
