+++ document-store/src/main/java/org/hypertrace/core/documentstore/query/transform/QueryTransformer.java
@@ -1,7 +1,7 @@
 package org.hypertrace.core.documentstore.query.transform;
 
-import org.hypertrace.core.documentstore.query.QueryInternal;
+import org.hypertrace.core.documentstore.query.Query;
 
 public interface QueryTransformer {
-  QueryInternal transform(final QueryInternal query);
+  Query transform(final Query query);
 }
