+++ document-store/src/main/java/org/hypertrace/core/documentstore/query/transform/TransformedQueryBuilder.java
@@ -1,11 +1,10 @@
 package org.hypertrace.core.documentstore.query.transform;
 
-import org.hypertrace.core.documentstore.query.QueryBuilder;
-import org.hypertrace.core.documentstore.query.QueryInternal;
+import org.hypertrace.core.documentstore.query.Query;
 
-public final class TransformedQueryBuilder extends QueryBuilder {
+public final class TransformedQueryBuilder extends Query.QueryBuilder {
 
-  public TransformedQueryBuilder(final QueryInternal query) {
+  public TransformedQueryBuilder(final Query query) {
     copySelections(query);
     copyFilter(query);
 
@@ -16,30 +15,30 @@ public final class TransformedQueryBuilder extends QueryBuilder {
     copyPagination(query);
   }
 
-  private void copySelections(final QueryInternal query) {
+  private void copySelections(final Query query) {
     // Iterate through elements to ensure deep-copy
     query.getSelections().forEach(this::addSelection);
   }
 
-  private void copyFilter(final QueryInternal query) {
+  private void copyFilter(final Query query) {
     query.getFilter().ifPresent(this::setFilter);
   }
 
-  private void copyAggregations(final QueryInternal query) {
+  private void copyAggregations(final Query query) {
     // Iterate through elements to ensure deep-copy
     query.getAggregations().forEach(this::addAggregation);
   }
 
-  private void copyAggregationFilter(final QueryInternal query) {
+  private void copyAggregationFilter(final Query query) {
     query.getAggregationFilter().ifPresent(this::setAggregationFilter);
   }
 
-  private void copySorts(final QueryInternal query) {
+  private void copySorts(final Query query) {
     // Iterate through elements to ensure deep-copy
     query.getSorts().forEach(this::addSort);
   }
 
-  private void copyPagination(final QueryInternal query) {
+  private void copyPagination(final Query query) {
     query.getPagination().ifPresent(this::setPagination);
   }
 }
